Small changes Added some useful methods to the SpriteSheet class IUpdatable and IDrawable interface are now explicitly functional The camera is now translating the view by double, not by int

+++ Sogomn engine/src/de/sogomn/engine/util/Camera.java
@@ -42,11 +42,11 @@ public final class Camera implements IUpdatable {
 	}
 	
 	public void apply(final Graphics2D g) {
-		g.translate(-(int)x, -(int)y);
+		g.translate(-x, -y);
 	}
 	
 	public void revert(final Graphics2D g) {
-		g.translate((int)x, (int)y);
+		g.translate(x, y);
 	}
 	
 	/**
